backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.2
flask-pymongo==2.3.0
marshmallow==3.21.1
python-dateutil==2.9.0
flask-jwt-extended==4.6.0
mypy==1.7.1
types-flask==1.1.6
types-pymongo==4.6.0.20240106
types-flask-cors==3.0.10.20240310
types-python-dateutil==2.8.19.20240316